requirements.txt
fastapi
pydantic
uvicorn[standard]
pyjwt
python-jose[cryptography] 
passlib[bcrypt]
python-multipart